backend/pyproject.toml
[project]
name = "voice-avatar-backend"
version = "0.1.0"
description = "VoiceLive + Avatar backend service"
requires-python = ">=3.11"
dependencies = [
    "azure-ai-voicelive>=1.0.0b1",
    "azure-identity>=1.19.0",
    "azure-search-documents>=11.4.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "websockets>=14.0",
    "python-dotenv>=1.0.0",
    "aiohttp>=3.13.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
